requirements.txt
aiohttp==3.6.2
altgraph==0.17
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.1
dnspython==2.0.0
future==0.18.2
idna==2.10
multidict==4.7.6
pefile==2019.4.18
PyInstaller==3.6
pymongo==3.11.2
pywin32-ctypes==0.2.0
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
